got rid of gui server based implementation
