+++ pylabnet/launchers/fiber_coupling_front.py
@@ -1,11 +1,11 @@
 from pylabnet.launchers.launcher import Launcher
 from pylabnet.launchers.servers import thorlabs_pm320e_front
-from pylabnet.scripts.fiber_coupling import power_monitor_direct
+from pylabnet.scripts.fiber_coupling import power_monitor
 
 def main():
 
     launcher = Launcher(
-        script=[power_monitor_direct],
+        script=[power_monitor],
         server_req=[thorlabs_pm320e_front],
         gui_req=[None],
         params=[None],
