+++ pylabnet/scripts/fiber_coupling/power_monitor.py
@@ -1,11 +1,15 @@
 import numpy as np
 from si_prefix import split, prefix
+import socket
 import time
 
+from pylabnet.gui.pyqt.external_gui import Window
 from pylabnet.utils.logging.logger import LogHandler
-from pylabnet.gui.pyqt.gui_handler import GUIHandler
-from pylabnet.utils.helper_methods import generate_widgets, unpack_launcher, find_client, load_config
+import pyqtgraph as pg
+from pylabnet.utils.helper_methods import generate_widgets, unpack_launcher, find_client, load_config, get_gui_widgets
 
+# Time between power meter calls to prevent crashes
+BUFFER = 5e-3
 
 class Monitor:
     RANGE_LIST = [
@@ -13,7 +17,7 @@ class Monitor:
         'R10MW', 'R100MW', 'R1W', 'R10W', 'R100W', 'R1KW'
     ]
 
-    def __init__(self, pm_clients, gui_client, logger=None, calibration=None, name=None):
+    def __init__(self, pm_client, gui='fiber_coupling', logger=None, calibration=None, name=None, port=None):
         """ Instantiates a monitor for 2-ch power meter with GUI
 
         :param pm_clients: (client, list of clients) clients of power meter
@@ -25,42 +29,82 @@ class Monitor:
 
 
         self.log = LogHandler(logger)
-        self.gui = GUIHandler(gui_client=gui_client, logger_client=self.log)
+        self.gui =Window(
+            gui_template=gui,
+            host=socket.gethostbyname(socket.gethostname()),
+            port=port
+        )
+
+        self.gui.apply_stylesheet()
         self.wavelength = []
         self.calibration = calibration
         self.name = name
         self.ir_index, self.rr_index = [], []
-        if isinstance(pm_clients, list):
-            self.pm = pm_clients
-        else:
-            self.pm = [pm_clients]
-
+        self.pm = pm_client
         self.running = False
+        self.num_plots = 3
+
+        # Get all GUI widgets
+        self.widgets = get_gui_widgets(
+            self.gui,
+            graph_widget=self.num_plots,
+            number_widget=4,
+            label_widget=2,
+            name_label=1,
+            combo_widget=2
+        )
+
         self._initialize_gui()
 
     def sync_settings(self):
         """ Pulls current settings from PM and sets them to GUI """
 
-        for channel, pm in enumerate(self.pm):
+        # Configure wavelength
+        self.wavelength = self.pm.get_wavelength(1)
 
-            # Configure wavelength
-            self.wavelength.append(pm.get_wavelength(1))
-            try:
-                pm.set_wavelength(2, self.wavelength[channel])
-            except:
-                self.log.warn('Failed to set the wavelength to channel 2')
-            self.gui.activate_scalar(scalar_label=f'wavelength_{channel}')
-            self.gui.set_scalar(
-                value=self.wavelength[channel],
-                scalar_label=f'wavelength_{channel}'
-            )
-            self.gui.deactivate_scalar(scalar_label=f'wavelength_{channel}')
+        self.widgets['number_widget'][-1].setValue(
+            self.wavelength
+        )
 
-            # Configure Range to be Auto
-            pm.set_range(1, self.RANGE_LIST[0])
-            pm.set_range(2, self.RANGE_LIST[0])
-            self.ir_index.append(0)
-            self.rr_index.append(0)
+        # Configure Range to be Auto
+        self.pm.set_range(1, self.RANGE_LIST[0])
+        self.pm.set_range(2, self.RANGE_LIST[0])
+        self.ir_index = 0
+        self.rr_index = 0
+
+
+        # Connect wavelength change action.
+        self.widgets['number_widget'][-1].valueChanged.connect(self._update_wavelength)
+
+        # Connect range change.
+        self.widgets['combo_widget'][0].currentIndexChanged.connect(lambda: self._update_range(0))
+        self.widgets['combo_widget'][1].currentIndexChanged.connect(lambda: self._update_range(1))
+
+    def _update_wavelength(self):
+        """ Updates wavelength of pm to WL of GUI"""
+
+        gui_wl = self.widgets['number_widget'][-1].value()
+
+        if self.wavelength != gui_wl:
+            self.wavelength = gui_wl
+            self.pm.set_wavelength(1, self.wavelength)
+            self.pm.set_wavelength(2, self.wavelength)
+
+
+    def _update_range(self, channel):
+        """ Update range settings if combobox has been changed."""
+
+        range_index = self.widgets['combo_widget'][channel].currentIndex()
+
+        if channel == 0:
+            if self.ir_index != range_index:
+                self.ir_index = range_index
+                self.pm.set_range(1, self.RANGE_LIST[self.ir_index])
+        elif channel == 1:
+             if self.rr_index != range_index:
+                self.rr_index = range_index
+                self.pm.set_range(2, self.RANGE_LIST[self.rr_index])
+
 
     def update_settings(self, channel=0):
         """ Checks GUI for settings updates and implements
@@ -68,7 +112,8 @@ class Monitor:
         :param channel: (int) channel of power meter to use
         """
 
-        gui_wl = self.gui.get_scalar(f'wavelength_{channel}')
+        gui_wl = self.widgets['number_widget_4'].value()
+
         if self.wavelength[channel] != gui_wl:
             self.wavelength[channel] = gui_wl
             self.pm[channel].set_wavelength(1, self.wavelength[channel])
@@ -85,18 +130,23 @@ class Monitor:
             self.pm[channel].set_range(2*channel+2, self.RANGE_LIST[self.rr_index[channel]])
 
     def run(self):
-        """ Runs the power monitor """
-
+        # Continuously update data until paused
         self.running = True
 
-        for channel, pm in enumerate(self.pm):
+        while self.running:
+            time.sleep(BUFFER)
+            self._update_output()
+            self.gui.force_update()
+
+    def _update_output(self):
+            """ Runs the power monitor """
 
             # Check for/implement changes to settings
-            self.update_settings(channel)
+            #self.update_settings(0)
 
             # Get all current values
             try:
-                p_in = pm.get_power(1)
+                p_in = self.pm.get_power(1)
                 split_in = split(p_in)
 
             # Handle zero error
@@ -104,13 +154,13 @@ class Monitor:
                 p_in = 0
                 split_in = (0, 0)
             try:
-                p_ref = pm.get_power(2)
+                p_ref = self.pm.get_power(2)
                 split_ref = split(p_ref)
             except OverflowError:
                 p_ref = 0
                 split_ref = (0, 0)
             try:
-                efficiency = np.sqrt(p_ref/(p_in*self.calibration[channel]))
+                efficiency = np.sqrt(p_ref/(p_in*self.calibration[0]))
             except ZeroDivisionError:
                 efficiency = 0
             values = [p_in, p_ref, efficiency]
@@ -121,114 +171,39 @@ class Monitor:
             formatted_values = [split_in[0], split_ref[0], efficiency]
             value_prefixes =  [prefix(split_val[1]) for split_val in [split_in, split_ref]]
 
-            plot_label_list = [
-                f'Input {channel+1}',
-                f'Reflection {channel+1}',
-                f'Coupling Efficiency {channel+1}'
-            ]
-            number_label_list = [
-                f'input_power_{channel}',
-                f'reflection_power_{channel}',
-                f'coupling_{channel}'
-            ]
-
             # Update GUI
-            for plot_no, plot in enumerate(plot_label_list):
-                self.gui.set_scalar(formatted_values[plot_no], number_label_list[plot_no])
-                self.plots[channel][plot_no] = np.append(self.plots[channel][plot_no][1:], values[plot_no])
-                self.gui.set_curve_data(
-                    data=self.plots[channel][plot_no],
-                    plot_label=plot,
-                    curve_label=plot,
-                )
-                if plot_no < 2:
-                    self.gui.set_label(text=f'{value_prefixes[plot_no]}W', label_label=self.labels[plot_no])
+            for plot_no in range(self.num_plots):
+                # Update Number
+                self.widgets['number_widget'][plot_no].setValue(formatted_values[plot_no])
 
-        self.running = False
+                # Update Curve
+                self.plotdata[plot_no] = np.append(self.plotdata[plot_no][1:], values[plot_no])
+                self.widgets[f'curve_{plot_no}'].setData(self.plotdata[plot_no])
+
+                if plot_no < 2:
+                    self.widgets["label_widget"][plot_no].setText(f'{value_prefixes[plot_no]}W')
 
     def _initialize_gui(self):
         """ Instantiates GUI by assigning widgets """
 
-        self.graphs, self.legends, self.numbers, self.labels, self.combos = generate_widgets(
-            dict(graph_widget=3, legend_widget=3, number_widget=4, label_widget=2, combo_widget=2)
-        )
-        self.plots = []
+        # Store plot data
+        self.plotdata =[np.zeros(1000) for i in range(self.num_plots)]
 
-        for channel in range(len(self.pm)):
-
-            # Graphs
-            channel_plots = []
-            plot_label_list = [
-                f'Input {channel+1}',
-                f'Reflection {channel+1}',
-                f'Coupling Efficiency {channel+1}'
-            ]
-            for index, label in enumerate(plot_label_list):
-                self.gui.assign_plot(
-                    plot_widget=self.graphs[index],
-                    plot_label=label,
-                    legend_widget=self.legends[index]
-                )
-                self.gui.assign_curve(
-                    plot_label=label,
-                    curve_label=label
-                )
-                channel_plots.append(np.zeros(1000))
-            self.plots.append(channel_plots)
-
-            # Numbers
-            number_label_list = [
-                f'input_power_{channel}',
-                f'reflection_power_{channel}',
-                f'coupling_{channel}'
-            ]
-            for index, label in enumerate(number_label_list):
-                self.gui.assign_scalar(
-                    scalar_widget=self.numbers[index],
-                    scalar_label=label
-                )
-                last_index = index
-
-            self.gui.assign_scalar(
-                scalar_widget=self.numbers[last_index+1],
-                scalar_label=f'wavelength_{channel}'
+        for plot_no in range(self.num_plots):
+           # Create a curve and store the widget in our dictionary
+            self.widgets[f'curve_{plot_no}'] = self.widgets['graph_widget'][plot_no].plot(
+                pen=pg.mkPen(color=self.gui.COLOR_LIST[0])
             )
-
-            # Assign prefix labels
-            for label in self.labels:
-                self.gui.assign_label(
-                    label_widget=label,
-                    label_label=label
-                )
-
-            self.gui.assign_container(
-                container_widget=self.combos[0],
-                container_label=f'ir_{channel}'
-            )
-            self.gui.assign_container(
-                container_widget=self.combos[1],
-                container_label=f'rr_{channel}'
-            )
-
-            # Assign name to name_label and set value.
-            self.gui.assign_label(
-                    label_widget='name_label',
-                    label_label='name_label'
-            )
-
-            self.gui.set_label(self.name, label_label='name_label')
 
 
 def launch(**kwargs):
     """ Launches the full fiber controll + GUI script """
 
     # Unpack and assign parameters
-    logger, loghost, logport, clients, guis, params = unpack_launcher(**kwargs)
-    pm_client = find_client(logger, clients, 'thorlabs_pm320e')
-    gui_client = guis['fiber_coupling']
+    logger, loghost, logport, clients, _, params = unpack_launcher(**kwargs)
+    pm_client = find_client(logger, clients, 'thorlabs_pm320e_front')
 
     # Unpack settings
-
     settings = load_config(
         kwargs['config'],
         logger=logger
@@ -239,19 +214,18 @@ def launch(**kwargs):
 
     # Instantiate controller
     control = Monitor(
-        pm_clients=[pm_client],
-        gui_client=gui_client,
+        pm_client=pm_client,
         logger=logger,
         calibration=calibration,
-        name=name
+        name=name,
+        port=logport
     )
 
     time.sleep(2)
     control.sync_settings()
-    while True:
 
-        if not control.gui.is_paused:
-            control.run()
+
+    control.run()
 
     # Mitigate warnings about unused variables
     if loghost and logport and params:
